requirements.txt
braintree==4.4.0
Django==3.1.1
python-dotenv==0.14.0
django-imagekit==4.0.2
Pillow==8.0.1